feat: 커리큘럼 기록 컴포넌트 분리
