+++ app/user/[id]/history/_components/LoadingState.tsx
@@ -1,0 +1,9 @@
+export default function LoadingState() {
+  return (
+    <div className='flex items-center justify-center py-32'>
+      <span className='material-symbols-outlined animate-spin text-5xl text-primary'>
+        progress_activity
+      </span>
+    </div>
+  );
+}
